refactor: auction  type에 대한 price 입력부분 컴포넌트 분리
